# ayunis-core-frontend/src/shared/api/openapi-schema.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/models/available": {
      "get": {
        "operationId": "ModelsController_getAvailableModelsWithConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved all available models",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelWithConfigResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all available models",
        "tags": [
          "models"
        ]
      }
    },
    "/models/permitted": {
      "post": {
        "operationId": "ModelsController_createPermittedModel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermittedModelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created a permitted model"
          },
          "400": {
            "description": "Invalid model input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a permitted model",
        "tags": [
          "models"
        ]
      },
      "get": {
        "operationId": "ModelsController_getPermittedModels",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved all permitted models",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermittedModelResponseDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all permitted models",
        "tags": [
          "models"
        ]
      }
    },
    "/models/permitted/{id}": {
      "delete": {
        "operationId": "ModelsController_deletePermittedModel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted a permitted model"
          }
        },
        "summary": "Delete a permitted model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/default": {
      "get": {
        "description": "Returns the applicable default model with fallback hierarchy: User default → Org default → First available model",
        "operationId": "ModelsController_getEffectiveDefaultModel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved the effective default model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermittedModelResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No default model found"
          }
        },
        "summary": "Get the effective default model for the user",
        "tags": [
          "models"
        ]
      }
    },
    "/models/org/default": {
      "get": {
        "description": "Returns only the organization's specific default model. Returns null if no org default is set.",
        "operationId": "ModelsController_getOrgSpecificDefaultModel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved the organization default model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermittedModelResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No organization default model found"
          }
        },
        "summary": "Get the organization default model",
        "tags": [
          "models"
        ]
      },
      "put": {
        "description": "Sets the specified permitted model as the organization default. If a default already exists, it will be updated to the new model.",
        "operationId": "ModelsController_manageOrgDefaultModel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetOrgDefaultModelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully set or updated the organization default model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermittedModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "404": {
            "description": "Permitted model not found"
          }
        },
        "summary": "Set or update the organization default model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/user/default": {
      "get": {
        "description": "Returns the user's personal default model only. Returns null if no user-specific default is set.",
        "operationId": "ModelsController_getUserSpecificDefaultModel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved the user-specific default model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermittedModelResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No user-specific default model found"
          }
        },
        "summary": "Get the user-specific default model",
        "tags": [
          "models"
        ]
      },
      "put": {
        "description": "Sets the specified permitted model as the user default. If a default already exists, it will be updated to the new model.",
        "operationId": "ModelsController_manageUserDefaultModel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserDefaultModelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully set or updated the user default model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermittedModelResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "404": {
            "description": "Permitted model not found"
          }
        },
        "summary": "Set or update the user default model",
        "tags": [
          "models"
        ]
      },
      "delete": {
        "operationId": "ModelsController_deleteUserDefaultModel",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Successfully deleted the user default model"
          },
          "404": {
            "description": "User default model not found"
          }
        },
        "summary": "Delete the user default model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/inference": {
      "post": {
        "operationId": "ModelsController_inference",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully triggered inference and returned model response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferenceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid inference request payload"
          },
          "404": {
            "description": "Model not found"
          },
          "500": {
            "description": "Internal server error or model provider error"
          }
        },
        "summary": "Trigger inference",
        "tags": [
          "models"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ayunis Core API",
    "description": "The Ayunis Core API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ModelWithConfigResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the model. Null if the model is not permitted.",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "The name of the model"
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "mistral",
              "microsoft",
              "ollama"
            ],
            "description": "The provider of the model"
          },
          "displayName": {
            "type": "string",
            "description": "The display name of the model"
          },
          "canStream": {
            "type": "boolean",
            "description": "Whether the model can stream"
          },
          "isReasoning": {
            "type": "boolean",
            "description": "Whether the model can reason"
          },
          "isPermitted": {
            "type": "boolean",
            "description": "Whether the model is permitted to be used"
          },
          "isDefault": {
            "type": "boolean",
            "description": "Whether the model is the default model"
          }
        },
        "required": [
          "id",
          "name",
          "provider",
          "displayName",
          "canStream",
          "isReasoning",
          "isPermitted",
          "isDefault"
        ]
      },
      "CreatePermittedModelDto": {
        "type": "object",
        "properties": {
          "modelName": {
            "type": "string",
            "description": "The name of the model",
            "example": "gpt-4"
          },
          "modelProvider": {
            "type": "string",
            "description": "The provider of the model",
            "example": "openai"
          }
        },
        "required": [
          "modelName",
          "modelProvider"
        ]
      },
      "PermittedModelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The id of the permitted model"
          },
          "name": {
            "type": "string",
            "description": "The name of the model"
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "mistral",
              "microsoft",
              "ollama"
            ],
            "description": "The provider of the model"
          },
          "displayName": {
            "type": "string",
            "description": "The display name of the model"
          },
          "canStream": {
            "type": "boolean",
            "description": "Whether the model can stream"
          },
          "isReasoning": {
            "type": "boolean",
            "description": "Whether the model can reason"
          }
        },
        "required": [
          "id",
          "name",
          "provider",
          "displayName",
          "canStream",
          "isReasoning"
        ]
      },
      "SetUserDefaultModelDto": {
        "type": "object",
        "properties": {
          "permittedModelId": {
            "type": "string",
            "description": "The ID of the permitted model to set as default",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "permittedModelId"
        ]
      },
      "SetOrgDefaultModelDto": {
        "type": "object",
        "properties": {
          "permittedModelId": {
            "type": "string",
            "description": "The ID of the permitted model to set as organization default",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "permittedModelId"
        ]
      },
      "InferenceResponse": {
        "type": "object",
        "properties": {}
      },
      "Message": {
        "type": "object",
        "properties": {}
      },
      "InferenceRequestDto": {
        "type": "object",
        "properties": {
          "modelName": {
            "type": "string",
            "description": "The name of the model to use for the inference",
            "example": "mistral-large-latest"
          },
          "modelProvider": {
            "type": "string",
            "description": "The provider of the model to use for the inference",
            "example": "mistral",
            "enum": [
              "openai",
              "anthropic",
              "mistral",
              "microsoft",
              "ollama"
            ]
          },
          "messages": {
            "description": "The messages to use for the inference",
            "example": [
              {
                "role": "system",
                "content": "You are a helpful AI assistant."
              },
              {
                "role": "user",
                "content": "Hello, what can you help me with today?"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "tools": {
            "description": "The tools to use for the inference",
            "example": [
              {
                "name": "weather_tool",
                "displayName": "Weather Tool",
                "description": "Get current weather information for a location",
                "parameters": {
                  "type": "object",
                  "properties": {
                    "location": {
                      "type": "string",
                      "description": "The city or location to get weather for"
                    }
                  },
                  "required": [
                    "location"
                  ]
                }
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "toolChoice": {
            "type": "string",
            "description": "The tool choice to use for the inference",
            "example": "auto",
            "enum": [
              "auto",
              "required"
            ]
          }
        },
        "required": [
          "modelName",
          "modelProvider",
          "messages"
        ]
      },
...
    }
  }
}